+++ evealert/__init__.py
@@ -1,4 +1,4 @@
 """Initialize the app"""
 
-__version__ = "0.9.0"
+__version__ = "1.0.0"
 __title__ = "EVE Alert"